feat: implement pagination for speaking days in course detail

+++ app/src/main/java/com/home/lexa/domain/repository/SpeakingDayRepository.kt
@@ -7,7 +7,7 @@ import com.home.lexa.domain.models.ShortParagraphSpeakingDayDto
 import com.home.lexa.domain.models.SpeakingDayPagination
 
 interface SpeakingDayRepository {
-    suspend fun getSpeakingDays(courseId: Long, nextOrder: Int?): Result<SpeakingDayPagination>
+    suspend fun getSpeakingDays(courseId: Long, nextOrder: Long?): Result<SpeakingDayPagination>
     suspend fun getParagraphSpeakingDay(speakingDayId: Long) : Result<ShortParagraphSpeakingDayDto?>
     suspend fun createSpeakingDay(request: CreateSpeakingDayRequest): Result<Long>
     suspend fun editSpeakingDay(speakingDayId: Long, request: EditSpeakingDayRequest): Result<Boolean>
